Support for Note ID, Poly and Mono modulation and more
This commit allows a host of surge, if the host provides it,
to thread a note id through the voice management so you can
handle start/end note with id messages (end coming from surge
back to the daw). It also allows all the parameters and macros
to separate modulation from value optionaly.

Given that, and given the update to clap-juce-extensions,
it re-implements the ::process loop inside SurgeSynthProcessor
to have an optional clap process loop which provides polyphonic
modulation.

Addresses #6137