Be more reslient in marking favorites
Favorites use a direct path match to mark as favorite but
if we change the location of the factory data folder that
means the db is there but the menu isnt. So for patches_factory
and patches_3rdparty, use a second check with a truncated path.

This is specially useful in windows XT1 -> XT11 when whe fixed
portable so the default data path changed

Closes #6177