LFO Randomness Issues (#6228)
in Deform mode 1, the first and only first modulator
value of SNH was always the same. Remediate this by
(effectively) running forward a few steps on attack.
Do the same for noise, but since noise has a peculiar
phase behaviour, that may be less handy.

Moreover, the seed strategy I used for independent RNGs
led to 'strpied' starting points making repepated attacks
seem to be non-random. Change to seed with the storage
audio thread rng32.