Start of CLAP non-destructive mods on integers and bools
This change means that non-destructive and polyphonic modulations
on integers and bools transmits through to surge. Surge, however,
does not read them reliably since they are placed in the mod matrix
and our DSP code has been indifferent to raw param val vs modulation
val access for ints and bools, as surge is not internally integer
and bool modulatable.

As such it starts to, but does not completely, address #6229

This commit makes oscillator octave modulatabple properly in clap