A couple of step seq out of bounds edit fixes (#6250)
First, when sweep dragging step seq triggers, these were constrained to just trigger rects. Update this so that we use event.pos.x so even if mouse leaves the trigger rects, the edits will apply.

Second, right mouse arrow drag in step seq would not be applied if mouse was released out of bounds. Fix this by applying the same point on line within rectangle constrant as done in paintStepSeq(). Abstract that into a method in SurgeGUIUtils.