Send Param Automation for Filter Subtypes
Paramter Automation messages were not sent because
basically filter subtypes are kinda screwed in how we
implement them both GUI wise and model wise. It's all a hack.
Anyway added an explicit send automation call and now when you
toggle through with clicks you end up getting the DAW showing changes.

My theory is this is the problem in #6264 but will wait for user
confirm before we close it.