Remove an ancient, unused load (#6347)
Back when surge was open sourced, it always loaded at least
one wavetable on startup. We have since replaced this with the
far more robust 'everBuilt' mechanism which populates all the
wavetables. But this little fragment still hung around. That meant
that the scene[0].osc[0] would load before everbuilt, but after wavetable
names were initialized, which caused the bug in #6248, of a name being
lost on init/save

Closes #6248