Revert partial modulatability of osc type
For a while I thought osc type was modulatable, but it wasn't
so I shut it off in can_be_nondestructively_modulated but forgot
to use the scene wide value again in the voice. This meant changing
osc types on playing voices led to inconsistent states. Revert
and deal with this some point in the far distant future.

Closes #6423