Defensive implementation of paramsFlush (#6480)
CLAP spec very strongly suggests (and i hope will soon unambivuously
say) that paramsFlush is for the case when processing is not active
but in CLAP 1.1.1 it is ambiguous and for a moment REAPER chose to
call paramsFlush when you called requestFlus even if processing was
going. That meant our defensive call to surge->process interfered
with audio. So do not call surge->process in paramsFlush if processing
is active, even though REAPER will rpobably stop calling me in this
case anyway.