Properly Scale the Waveshaper Effect
The Waveshaper Effect was mis-scaled relative to the in-voice
waveshapers so exhibited a different response. This was because
of the network of up and down samples and various other factors
but it made it seem pretty broken. Now it meets the consdtraint
that sin -> filterbank -> output and sin -> fxunit -> output
respond the same on a same shaper at same drive.

Closes #6565