Move Oscilloscope background to a sibling component of the display.
This will keep it from constantly re-rendering whenever the display
re-renders. For whatever reason, re-rendering that background with all
the lines consumes a ton of CPU. It can't be cached with Juce's image
caching either, since when a *child* component (the spectrogram)
updates, the parent gets invalidated and must repaint. So it must be a
sibling.

Interestingly, I didn't note any difference after making it a sibling
in terms of CPU usage when it was buffered to an image cache or not.
So I just turned that off.