Add an optional LFO onepole for FROM_LAST envelope mode
In envelope mode with high deform in FROM_LAST mode the
signal from the envelope is not continuous. This is basically
unsolvable analytically - the deform plus state is too hard -
and so leave it on by default but add an optional one pole to
smooth it out, which may be useful in some cases.

Expose configurations of that to the rack LFO

Addresses https://github.com/surge-synthesizer/surge-rack/issues/607