Fix WT-length-one clamp error
In recent versions we allowed wavetables of length one
but when doing so missed this clamp which resulted in
hi < lo. that is UB (but defacto clamped to low) but
that inversion triggers an assert on some configurations.

So (1) add a debug assert to our limit range and (2) use a max of
0 on the upper bound.

Closes #6702