An optional off-by-default Twist LPG Fix
See issue #6760 and https://github.com/surge-synthesizer/surge-rack/issues/766

The Twist LPG implementation used an incorrect block size so decayed too
quickly. There's probably a correct patch to plaits to not decay with
size 12 (around line 208 of voice.cc) but for now make it correctable
optionally so we can correct it in rack.