Invalidate all params etc... on patch change (#6780)
A patch change can change the name or value of any param.
So when it happens send the maximal updateHostDispolay (excluding
latency). Trigger this with a new patchChanged atomic on the synth
which we set to true at the end of the off-thread load. Results
in bitwig showing all correct param changes on patch change in vst3
and presumably other compliant hosts would do same.

Closes #6607