Properly zero out Oneshot I16 padding (#6785)
When building a one shot we pad with 3 extra tables.
Those tables get zero values in them. BUT the zero table
for the I16 mipmap has an FIR extension and our zeroing
didn't zero this out. This meant that if you load a wave
as oneshot (but not a wt which would be correctly streamed)
then save and unsave, you could get boogers in the end of the
last frame. Nothing w ecan do about people who streamed in the
past but this will fix it going forward.

Closes #6573