Step Sequencer RMB / Edit Part 1
This opens a menu on the RMB gesture on the step sequencer and
preserves the right drag with a timer; it links to documentation
properly and formats the item. The only thing we need to do now is
to pop upen the edit box and figure out what else we want on this
menu. Neither of those are easy so checkpointing along the way.

Menu pops from both mouse and accessible gestures.

Addesses #6516

foo