MSEG Constant Segments trigger properly (#6916)
MSEG Constant segments have an optimization to skip most of
the calculation which is great. But it avoices updating an internal
state about triggers meaning a segment would fire a trigger on
every sample if exactly constant. Update the evaluation time
state to make sure we know we are beyond the trigger.

Closes #6915