Don't send the Audio Latency message to stdout
Generally we send errors to stdout but the audio latency is more
an informational so add an option to not stdout things from
reportError and set it true in the processor when we turn on
latency. The dialog will still pop up with a don't warn again
in the UI once you open it though.