Osc->init ignored octave; commit changes sound.
SurgeVoice called osc->init with a pitch which ignored
the octave offset and so forth. This meant that the first block
of any oscillator with octave +3 would have a 3 octave immediate
shift.

This long lived bug (since inception) is audible in the String
and Modern oscillators which have a variety of memory and smoothing
in their handling of pitch.

This will subtly change the sound of oscillators with high octaves,
but will change it to be correct.

Closes #6985