Slightly tweak basic-blocks Params api
In pursuit of doing #6991 a few more value
measures need to get to value-to-string and modvalue-to-string
namely the temposync an dextended state. Rather than a confusing
list of bool args, do this with a little helper state object
which currently still results in an implementation which
ignores it, but will let me solidy the api so I can start
the SC and standalone juce client thingy port.