Alert has its own Focus Traverser
Alert doesnt' subclass Overlay. That's fine but as a result it participates
in main window focus. Change it so it has its own keyboard traverser.

This mostly solves the 'wander into other ui' problems but I still
get the patch browser in my keyboard list so its not complete. Merge
to make nightlies better for ally users tho

Addresses #7089