Turn off Midi PC Factory Backfill
We initially made the unused banks fill in order
with Factory. We decided this was probably too volatile
so just turned it off.

Doing so means in most cases midi program change does
nothing so also remove the menu item for ignore midi program
change.

Also push MIDI Programs to bottom of user list after a separator
and only show it if it has contents

Closes #576