Modern Oscillator Pitch: Lag -> Lipol
addresses #7224

The original version of Modern used the trailing three points
for the derivative, but didnt' recaclulate all three at once.
This meant it was a bit unstable under extreme pitch changes
so we used a lag/lpf on pitch and dphase. We subsequently changed
the algorithm to always calculate the trailing three points but
never undid that lag, leading to low porta high shift having an
audible sweep in modern, as reported in #7224.

This change fixes that and simply block smooths pitch shifts.