Slider-Mouse-Unhide error; Flanger temposync fix
The horitonzal slider positioned the mouse subtly
incorrectly when it unhid, placing it at the top of the
slider box, which in the tightly packed FX section meant mis-clicks
were easy. Closes #7232

Also the temposycn scale on teh flanger rate was wrong, but
fix it using the API I just added to do same in Phaser.
Closes #7233