MONO/MONO FP Gated Porta Fix. Behaviour Change
In Mono and Mono FP the portamento would work incorrectly
on a held release with a partial slide. that is, set portamento
to a long time, Press C, press G wait until you get to about E
in the slide, release the G. The porta-slide-back would start from
G.

This makes it start from E with the appropriate phasing and time.

It is a change in bheaviour but a correct behaviour so there
is no bailout option right now.

Closes #7301