Integrate Superparasites Clouds Firmware (off by default)
This commit implements the superparasites firmware as an
option for clouds. The firmware comes from patrick dowling
at https://github.com/patrickdowling/superparasites

In this commit, the firmware is activatable at build time
but is still off by default, so the synth is unchanged
unless you opt in in libs/eurorack/CMakeLists with the option.

Closes https://github.com/patrickdowling/superparasites/issues/6
Addresses #7359