Tuning Editor Keyboard can also play notes
Forever and a day that keyboard table layout of frequencies
has been sitting there, wanting to be pressed to play a note.
But in the time we've had it noone thought of it until VincyZed
suggested it.

Anyway bit more of a refactor than I thought since the juce
tablelistboxmodel thing is very opinionated so needed to factor
to a custom component etc... but also usefully adds a ui safe
play and release note method to SurgeGUIEditor

Closes #7365