Drain the osc ring buf without audio
When audio isn't running the osc ring buf just fills up.
Like we do with other things use the audio processing active
bool to drain on the osc thread. There's a tiny race condition if
you are processing an osc message while you turn on audio but
leave that edge-case-of-an-edge-case in a comment

Closes #7445