Patch Selector Favorites/Find button Accessible; Focusable
The favorites and find button were funky - drawn in place.
They are still a bit funky - drawn in place with a focusable
accessible overlay - but that's just to avoid reworking the draw
code. This makes them accessible, focusable, and so on. Moreover
the accessible name of the favorites button indicates whether a
patch is a favorite or not.

Closes #7399