Different recovery path for hosed %DOCUMENTS% (#7552)
* Different recovery path for hosed %DOCUMENTS%

Seems OneDrive can hose down %DOCUMENTS% on windows enough
that more people were getting the error that the folder cant
resolve. This sucks and those systems are misconfigured.

But with these changes we actually can start surge with a clear
error that says saving and the like wont work. The synth kinda works
also

f

* Oops - windows typo