Start and stop the NSEventLoop in the CLI for OSC (#7567)
OSC uses the NSEvntLoop on MacOS to send messages in
some cases. Whereas on linux this is explicitly
driven by a client thread on mac you need to start and
stop the system. Use the internals fo the juce mac application
to do so, and a custom function to shut down cleanly.