Clean up channel keyboard state from latch mode
The channel state was left in a corrupted state when not latching.
This meant latch -> mono -> key -> release could trigger the latch
voice if the key you pressed was not the latch voice. There was
even a comment to clean up, but we ended up cleaning up by
terminating the entire scene in latch (more appropriately)

So this also udpates the keybed state to avoid that
voice edge case

Closes #7628