Github Actions Release Stage (#7741)
Move the nighly build to github actions. Have framework
for the release build but leave that on azure for a moment still.
Adjust all the various things which make surge special like
the version and website, but use as much of the shared actions
stuff as possible (which is most of it!)