Improve patch search result list behavior
Previously, if we reopened search results that already had one result focused/loaded, pressing the down arrow key would select the first entry in the results list.
This is not really intuitive, it would be expected to move to the next entry if we press the down arrow, or the prior entry if we press the up arrow.

This PR makes it so!