Reduce excessive abbreviation in SurgeSynthEditor
Some examples: sge -> surgeEditor, p-> parameter.

As much as it can be "obvious". I am new to the code base and it is
a difficult to get the gist of basic things because I have to
backtrace and remember every abbreviation since they are not even obvious.

I understand the spartan programming mindset, but the Developer Guide
encouraged "campground" commits like this.

I hope this makes Surge better even if it is only a small change.